+++ __tests__app__/app/~rs/component.map.js
@@ -1,4 +1,0 @@
-import __9b09814e6bb3157f4bbc9d33a0112690__ from '../test'
-import __5ed5026db447e578254c5f81cc0e05f3__ from '../socket-status'
-import __9a917bcd1d8eca2508e2519c045e747f__ from '../generated-component'
-export default [{elementId: 'test', component: __9b09814e6bb3157f4bbc9d33a0112690__},{elementId: 'socket-status', component: __5ed5026db447e578254c5f81cc0e05f3__},{elementId: 'generated', component: __9a917bcd1d8eca2508e2519c045e747f__}];