+++ __tests__app__/app/~rs/router.map.js
@@ -1,8 +1,0 @@
-import __e2fa272b0b69291f2916ccf63f55d0e8__ from '../home';
-import __e58be13276af7ff461a5da99152c5a09__ from '../about';
-import __686be7eb0614db5981c9a6d6e4254ee7__ from '../socket';
-import __308628e6ff7adac9a6627b99d106e684__ from '../decorators';
-import __0990c62b8eafeaa5ce069867687d7774__ from '../models';
-import __d516d3dbe748fa4f0813f1ceea9d6f6a__ from '../home';
-import __0de4b02de3fa7f24c7c87d4746818b50__ from '../generated';
-export default [{spec: '/', component: __e2fa272b0b69291f2916ccf63f55d0e8__, title: 'Home'},{spec: '/about', component: __e58be13276af7ff461a5da99152c5a09__, title: 'About'},{spec: '/socket', component: __686be7eb0614db5981c9a6d6e4254ee7__, title: 'Socket'},{spec: '/decorators', component: __308628e6ff7adac9a6627b99d106e684__, title: 'Decorators'},{spec: '/models', component: __0990c62b8eafeaa5ce069867687d7774__, title: 'Models'},{spec: '/layout-test', component: __d516d3dbe748fa4f0813f1ceea9d6f6a__, title: 'Layout test'},{spec: '/generated-page', component: __0de4b02de3fa7f24c7c87d4746818b50__, title: 'Generated page'}];