+++ __tests__app__/app/~rs/socket.map.js
@@ -1,1 +1,0 @@
-export default ['socket.test','socket.file','user.get','user.getPromise','user.getAsyncError','user.getSyncError','user.getVoidPromise','user.getPayloadedError'];